Updated from Kana ------------------------------------------------------------------------ r1562 | KanaSugimoto | 2009-04-08 03:20:34 -0600 (Wed, 08 Apr 2009) | 21 lines
       New Development: No
            JIRA Issue: Yes (CAS-1298)
      Ready to Release: Yes
     Interface Changes: No
What Interface Changed: 
         Test Programs: run sdplot() with wrong 'sprange' or 'flrange' selections
  Put in Release Notes: No
             Module(s): tools and tasks which calls sd.plotter.plot()
           Description: 
Methods plot() and _slice_indeces() are modified to fix bugs in ASAP plotter. 
The method _slice_indeces() is modified to avoid index referencing error and 
to return precise 'slice indices'. 
The method plot() is modified to avoid invalid index referencing when setting 
plot range. The y-range of the plot will be set to the default plot range of 
Matplotlib, i.e., [0,1], if flrange is not set and the selected sprange is out 
of range. 




git-svn-id: https://svn.cv.nrao.edu/svn/casa/v2/stable@7702 ac9c0da5-131f-0410-9009-a9c195f183e6