New Development: No             JIRA Issue: Yes (CAS-1076)       Ready to Release: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s): Functions calling SDGrid::getIndex()            Description: Function SDGrid::getIndex() is modified. The boundary value is added to meet condition in backward search of pointing information to avoid index not found error when integration time is divisible by 2.
Function SDGrid::getIndex() returns pointing information for a 
specified time (in MAIN table) by looking for the time range of 
each poinings (calculated from POINTING sub-table) which includes 
the specified time. The values in time collumn in MAIN table denote 
the time centroid of the scan while those in the POINTING table do 
the beginning of integration. So, in practice, the method checkes 
if the deviation of time between the specified one (time centroid) 
and one in pointing table (start time) is smaller than the half of 
integration time in each pointing. Since the previous code doesn't 
include the boundary value for the judgement, it sometimes failed 
to find pointing information in the case integration time was 
divisible by 2. 
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/v2/active/code/synthesis@8013 ac9c0da5-131f-0410-9009-a9c195f183e6