New Development: Yes             JIRA Issue: Yes - CAS 330,241,1382       Ready to Release: No      Interface Changes: Yes What Interface Changed: Added new class MSPolnParse.  Added global             functions in MSPolnGram to drive the MSPolnParse.  This is             the interface used in MSSelection class.          Test Programs: List test programs   Put in Release Notes: Yes              Module(s): Module Names change impacts.            Description: Describe your changes here...
Added the polarization selection expression parsing class
MSPolnParse.  The global functions to drive the parsing are in
MSPolnGram.{cc,h}.  This is the interface used in the MSSelection
class.  Since the parsing for this expression is done by-hand
(i.e. not via program generation tools (flex/bison)), the driver
functions need not be global (in MSPolnGram).  But keeping it that was
to give a reasonably uniform interface for parsing all expression.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/casacore/alma-evla@8479 ac9c0da5-131f-0410-9009-a9c195f183e6