New Development: Yes             JIRA Issue: No       Ready to Release: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes              Module(s): Module Names change impacts.            Description: Describe your changes here...
Changes to allow replacing the ChannelStep by a caller defined default
value.  If the channel step is omitted as part of the CHAN selection
expression, the default value for the step can be given as a parameter
to getChanList() call.  The default value of the defaultStep is 1
leaving the current behaviour unchanged.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/casacore/alma-evla@8903 ac9c0da5-131f-0410-9009-a9c195f183e6