New Development: No             JIRA Issue: Yes CAS-1823       Ready to Release: Yes      Interface Changes: Yes What Interface Changed: Added varlist parameter in sdmath task.          Test Programs: see below   Put in Release Notes: Yes              Module(s): Module Names change impacts.            Description: Describe your changes here...
I have added varlist parameter that should be used combining with expr 
parameter. If you define varlist dictionary, its key can be used in 
the expression as a variable, and corresponding value is substituted 
when expression is executed. Value for varlist dictionary can be 
float, (1D or 2D) list, or string which is a filename of scantable. 




git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@10143 ac9c0da5-131f-0410-9009-a9c195f183e6