New Development: No             JIRA Issue: Yes CAS-1823       Ready to Release: Yes      Interface Changes: Yes What Interface Changed: The factor accepts string value in sdscale task          Test Programs: List test programs   Put in Release Notes: Yes/No              Module(s): Module Names change impacts.            Description: Describe your changes here...
The sdscale task allows to set factor as filename. The file should consist
of list of space-separated scaling factor with adequate number of rows and 
columns (should satisfy conformity with input scantable).
The sdmath task also allows to set operand as filename whose content satisfies 
the same condition as above.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@10171 ac9c0da5-131f-0410-9009-a9c195f183e6