New Development: No             JIRA Issue: No       Ready to Release: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s): evla tasking            Description: changes from Steve
Attached are revised versions of importevla task files to be put into active.
These add the "flagzero" (default=T) and "shadow" (default=F) params with
relevant sub-params.

When these get into active, let Joe and Gustaaf know so they can test.  If it
works well (I only tested on one dataset), then push to casapy-test so the
archive can pick this up.

Note that the default behavior of importevla will then be:
  apply online flags
  clip out RR and LL "zeroes" (but not RL,LR could add that), below 1E-8
  do NOT flag shadowed data (needs more testing)
thus it should be ok to average in time (except for the RL,LR issue).

Eventually want to change flagdata to allow extending flags (e.g. from the
clip on RR and LL) to all corrs, probably need a JIRA for that.


	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@11081 ac9c0da5-131f-0410-9009-a9c195f183e6