New Development: No             JIRA Issue: No         Ready for Test: Yes and probably stable      Interface Changes: Yes What Interface Changed:          Test Programs: importeval_regression   Put in Release Notes: Yes              Module(s): tasking, evla reduction            Description: From Steve Myers...
Attached is a new version of the task importevla, which combines
the 2 (or 4) separate calls to flagdata for flagzero=T into
a single one.  This cuts the time significantly >20% for the flagging
(more if flagpol=T).

This is just changing the call to flagdata inside the task slightly,
and seems to work ok in tests.  I'd like to get this into active for
further testing, and then probably soon into the archive version.
Could even go into release, I don't think there are bad implications,
but should test in active and/or test first.


	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@11635 ac9c0da5-131f-0410-9009-a9c195f183e6