New Development: No             JIRA Issue: Yes CAS-2136, 2137         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Describe your changes here...
1. Indent level changed

2. Create CORRECTED_DATA column if necessary
CORRECTED_DATA column is created only if calmode='baseline'.

3. plotlevel bug fix
Fixed a bug that plotter panel is displayed even if plotlevel=0.

4. output EPS filename (plotlevel<0)
Changed filename that depends on polarization so that the file 
is created for each polarization (only the last plot is saved to 
file before this fix so that previous plot is lost).

5. Other
Bug fix on the code that gets polarization information from POLARIZATION 
subtable.
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@11643 ac9c0da5-131f-0410-9009-a9c195f183e6