New Development: No             JIRA Issue: CAS-2317         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs: cvel_regression.py   Put in Release Notes: No              Module(s): scimath            Description: I found that in a special case for linear interpol with InterpolateArray1D, the flagging doesn't seem to behave correctly.
If the grid is aligned with the old grid such that the interpolation is trivial,
then the new flag is still determined from the two input flags of the "interpolated"
channels even though only one of the two channels actually contributes.
If one of the input channels is flagged, the output channel is as well.

In this way, a flag can "leak" to the next channel. 

This behaviour is now corrected.
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active/casacore@12187 ac9c0da5-131f-0410-9009-a9c195f183e6