New Development: Yes             JIRA Issue: CAS-1881 (plotms iteration)         Ready for Test: No      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s): plotms            Description:
"Iter" tab appears for multiplot.  Text on 
some controls turns red when user changes the
value.  Some controls don't do this yet.  

All controls (we hope) are connected to parameters in 
PMS_PP_Iteration, defined in PlotMSPlotParameterGroups.h, 
and should be usable in new code to be written 
by George and Daren.

Unknown if these params are available to casapy,
but probably.  Nothing related to Iteration or Multiplot
should be tested at this time.  Work is in progress,
could crash, authorized personnel only, and all that.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@12313 ac9c0da5-131f-0410-9009-a9c195f183e6