------------------------------------------------------------------------ r12363 | dirk.petry | 2010-08-19 10:23:49 -0600 (Thu, 19 Aug 2010) | 19 lines Changed paths:    M /active/casacore/casa/BasicSL/Complex.h
       New Development: No
            JIRA Issue: CAS-2381
        Ready for Test: Yes
     Interface Changes: Yes
What Interface Changed: Complex.h
         Test Programs: manual tests
  Put in Release Notes: Yes?
             Module(s): casacore/casa
           Description: Added inline functions to multiply
                        Double * Complex (in this order)
                        and Int * Complex. Complex * Double
                        and Complex * Int were already present.
                        Necessary, e.g., for InterpolateArray1D
			to work for <Double, Complex> .
			Sorry folks: need to recompile all of CASA.
	


------------------------------------------------------------------------
r12367 | wes.young | 2010-08-19 16:32:14 -0600 (Thu, 19 Aug 2010) | 20 lines
Changed paths:
   M /active/casacore/casa/IO/MMapfdIO.cc
   M /active/casacore/casa/IO/MMapfdIO.h


       New Development: No
            JIRA Issue: No
        Ready for Test: Yes
     Interface Changes: No
What Interface Changed: 
         Test Programs: regressions, especially g192 and ggtau
  Put in Release Notes: No
             Module(s): IO
           Description: 
Made a change to avoid constant mmap and munmap when growing a file.
I reverted to Jonas's previous version that trys to mmap the whole file into memory as
this got rid of some horrendous split and making back flag tables timings on the mac.
Then for extending the file I expand the file by a factor of 2 of the "FileSize" to 
reduce the number of unmap, extend, map calls which speeds things up on the mac so
filling is not so pokey. Once all the data are read in I shrink the file to the right
size. So we don't have bloat in things that use the TSM.
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/test@12384 ac9c0da5-131f-0410-9009-a9c195f183e6