New Development: Yes             JIRA Issue: (tbd)         Ready for Test: No      Interface Changes: Yes What Interface Changed: plotms          Test Programs:   Put in Release Notes: No              Module(s): plotms            Description: Ranges in axes tab are now easier to edit for date/time. What was a difficult to use custom widget is now just a simple text edit box, for both "from" and "to" values.
The color of the text changes according to its validity.
Correct date/time is black.   It's blue while being
written in, and red if there's an incorrect character.
Strange values such as the 33rd day of the month should
turn the text purple, but there seems to be a bug.

This check-in is for unofficial testing and comment
by invited users.   Further refinement is necessary,
and a proper source file location for some added code
needs to be decided upon.
 
This change applies to both X and Y axes in the Axes tab.

This work grew out of a discussion on CAS-2117.


               


git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@12479 ac9c0da5-131f-0410-9009-a9c195f183e6