New Development: Yes             JIRA Issue: -         Ready for Test: Yes      Interface Changes: Yes What Interface Changed: plotms          Test Programs:   Put in Release Notes: Yes              Module(s): plotms            Description:
Manual ranges in the axes tab are now easier to edit for 
date/time.  What was a difficult to use custom widget is 
now a simple text edit box,  for both "from" and "to" 
values, that users can paste text into, copy, and edit 
freely. 

This change applies to both X and Y axes in the Axes tab, 
and applies only for Manual range settings.

The color of the text changes according to its validity.
The user will normally see blue while entering a 
date/time.  When it includes everything from years to
minutes, it'll turn black.  Seconds are optional, and
assumed to be zero if absent.  

Typing errors will cause the text to turn red or purple.
Red indicates problems with superficial syntax, that the 
wrong kind of character was used e.g. 'w' where
a digit was expected.   Purple shows that the superficial
syntax is okay, but the value of one of the numbers
is nonsense, e.g. month 13, or 67 minutes.


New code was added to the RangeWidget source, but some of
it probably should be packaged better and kept elsewhere,
maybe associated with Time() in casacore, or put into 
casaqt.   An idea is to derive a DateTimeEdit control
from Qt's QLineEdit, to include this colorizing feature.

This work grew out of a discussion on CAS-2117, but is 
a separate modification from fixing that issue.


               


git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@12497 ac9c0da5-131f-0410-9009-a9c195f183e6