New Development: No             JIRA Issue: Yes CAS-2450         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes              Module(s): Module Names change impacts.            Description: Describe your changes here...
The equation used to compute the pixel co-ordinate on a uv-grid was
incorrect (and inconsistant with that used for gridding).  Fixing this
in VisImagingWeight::VisImagingWeight() (lines 69,70) fixes the
aliasing problem, most clearly seen in PSFs of E-W arrays where the
PSF sidelobes are coherent and strong far from the phase center.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@12796 ac9c0da5-131f-0410-9009-a9c195f183e6