New Development: Yes             JIRA Issue: CSV-2038         Ready for Test: Yes      Interface Changes: Yes What Interface Changed: casaplotms command line options          Test Programs:   Put in Release Notes: Yes              Module(s): plotms            Description: When plotms is run with the command line option --nopopups, it will show error and other messages in the status bar along the bottom of the main window, and write them to the log.
This is useful for running plotms automatically, where no 
one is attending the computer but plotms is command to show 
incoming data in near real time, for example as done by 
ALMA.   If a popup message were to appear, and no one 
is around to click "OK", plotms execution is blocked.
With --nopopups, execution continues.


There is one place in plotms where it asks
	"One or more threaded operations are not yet 
     complete!  Do you still want to quit?"
In --nopopups mode, the answer is assumed yes.  
  


git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@13067 ac9c0da5-131f-0410-9009-a9c195f183e6