New Development: No             JIRA Issue: Yes (CAS-2521)         Ready for Test: Yes      Interface Changes: Yes What Interface Changed: MSSelection::getChanList() has a new parameter             "sorted" which is defaulted to False.          Test Programs: List test programs   Put in Release Notes: Yes              Module(s): Module Names change impacts.            Description: Describe your changes here...
The output of MSSelection::getChanList() was incorrect for certain
type of expressions.  The bug was in copying the internally generated
table of SPW IDs and associated chan. ranges to the temporary
table. Fixed that.

Also added a parameter to MSS::getChanList() to control if the output
table be sorted by SPW IDs or not.  The default is "do not sort" since
that seems to be the path of least regression failures at this point.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@13073 ac9c0da5-131f-0410-9009-a9c195f183e6