New Development: Yes             JIRA Issue: No         Ready for Test: No      Interface Changes: Yes What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Describe your changes here...
Made several addtions and changes to make it a more respectivable
(i.e. usable) container class for holding the convolution function
pixels and other associated aux. info.  This is the object that now
gets pass around within the C++ code.  The pixel array is held as a
Array<Comple> which is explicitly used only for passing it down to the
FORTRAN code.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active/code/synthesis@13373 ac9c0da5-131f-0410-9009-a9c195f183e6