New Development: Yes             JIRA Issue: No         Ready for Test: No      Interface Changes: Yes What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Describe your changes here...
Made serval changes and bug fixes to plug it into the existing
FTMachine (AWProjectFT.cc). The goal is to make it such that the only
method required by the client code would be "locateConvFunction".
This class should optimally handle these signals
(locateConvFunction).  I.e. it should efficiently locate the function
if it is cached, not leak memory and manage the cache (memory and disk
caches) intenally.  None of this code should leak to the application
layer. 

This check-in represents the first cut at the above functionality.
This is now plugged into AWProjectFT which passes the relevant
regressions (ic2233_regression.py).
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active/code/synthesis@13374 ac9c0da5-131f-0410-9009-a9c195f183e6