New Development: No             JIRA Issue: Yes CAS-2609         Ready for Test: No      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Describe your changes here...
Moved CFCache and CFStore (for hold convolution functions) to the
FTMachine baseclass.  

Also made normalizeImage() a pure virtual method
of FTMachine.  This required impementation of this method in all
drived class.  Currently AW*FT machines have a non-trivial
implementation of this method.  In other derived class an exception is
thrown if this method is called.
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@13627 ac9c0da5-131f-0410-9009-a9c195f183e6