New Development: Yes             JIRA Issue: CAS-2440         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes:              Module(s): plotms casaqt genericplotter            Description: User may now turn on/off vertical and horizontal axes independently.   Before this fix, to have horizontal visible and vertical not meant getting neither.
The root problem was use of ! instead of ~ to flip
bits in a bitset (integer) indicating the visiblity
of scales for the each of the four sides.  

To clarify the intent of parts of the relevant code, 
a new typedef has been declared, PlotAxisBitset. 
Several methods in PlotCanvas (generic plotter) and
QPCanvas (casaqt) were changed to use this.  Some
#includes added to make this typedef available may
cause more recompilation than truly necessary.  Some
other details including formatting improvements,
dead code removal, warning shut-ups, etc were made in the
course of chasing this bug, with no changes in function.

One file, TBConstants, was giving annoying warnings 
due to signed/unsigned comparisons, unrelated to this bug.
This file happened to get caught in this commit.




git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@13687 ac9c0da5-131f-0410-9009-a9c195f183e6