New Development: Yes             JIRA Issue: No List JIRA ticket.         Ready for Test: Yes      Interface Changes: Yes What Interface Changed: see below          Test Programs: List test programs   Put in Release Notes: Yes/No              Module(s): Module Names change impacts.            Description: Describe your changes here...
Added a new method to imager to update residual while keeping the state of imageskymodel. But also now modified imager so that clean and restore releases the residual image it does not need to keep a lock. 
These are necessary for parallelization and sharing of residual images at different stages
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@13704 ac9c0da5-131f-0410-9009-a9c195f183e6