New Development: No             JIRA Issue: Yes CAS-2609         Ready for Test: No      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Describe your changes here...
Added a memCacheWt_p object to manage to average PBs.  For multi-term
Wide-band imaging and for mosaic imaging, this will be put to
non-trivial usage.

addToMemCache() not also takes the internal CFStoreCacheType object
for in-memory caching.

The image containing the conv. funcs. on the disk now have "miscInfo"
record which holds the support size, sampling and PA values.  These
are written to the disk in cacheConvFunction() method and read in
loadFromDisk() method.  With this, the need for holding these values
in the "aux.dat" file is gone (though it still has these values).

An internal avgPBReady_p flag is kept to indicate if avgPB was
read/computed/cached or not.  This is used in AWConvFunc and AW*
FTMachines to trigger computation of avgPB or not.

Added a locateConvFunction() method to also load the weight functions
from the disk cache or locate it in the mem-cache.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active/code/synthesis@13747 ac9c0da5-131f-0410-9009-a9c195f183e6