New Development: Yes             JIRA Issue: CAS-1971         Ready for Test: Yes      Interface Changes: Yes What Interface Changed: GUI: new button to subtract selection regions. API: Enum for actions, tools in generic plot.
         Test Programs: -
  Put in Release Notes: yes!
             Module(s): plotms, casaqt, genericplot
           Description: 

Added Region Subtraction feature to resolve JIRA
CAS-1971.  This entails changes to a couple .ui files,
addition of a .png image, and numerous sites of 
code changes in the source including mods to datatypes
and certain class methods defined in generic plot, casaqt 
and plotms classes.


USER INTERFACE CHANGES 
A new button appears next to 
the mark region '+' and delete all regions 
red-circle-slash buttons under the main 
plotting pane.  The new button shows 
a violet minus sign.   When it's down, the user 
drags a rectangle which will erase all existing 
selection rectangles contained entirely within.  

Note that the original JIRA ticket suggests using
a key, e.g. 'd' for delete, while holding the mouse
cursor inside a selection rectangle to be zapped.
This was not possible due to key press events 
not appearing in suitable places where appropriate 
actions could be taken.  Likewise it wasn't clear how
one could add a ctrl-Z "undo" feature.


REMAINING WORK TO DO
Need a way to draw the "zapper" rectangle in a
different line style while it is being dragged
("rubber band" mode) by the user.  To do this, PlotTool.cc 
needs access the QwtPlotPicker class.  This means adding an 
#include that won't compile.  The cmake build system needs 
to be informed that -I/usr/include/qwt should be given to gcc.

PlotSelectTool has a new member m_subLine, a linestyle
to draw when the user rubber-bands the subtraction 
box.  It's used in various places that didn't produce
the desired effect.  Such code may or may not be 
needed when the proper QwtPlotPicker stuff is in 
place.  Clean it out if it's not needed.
 


git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@13800 ac9c0da5-131f-0410-9009-a9c195f183e6