New Development: No             JIRA Issue:         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: Yes              Module(s): importasdm.            Description: Now fills the content of column NOISE_CAL in MS CALDEVICE with the content of the column coupledNoiseCal in ASDM::CalDevice if it exists, otherwise fills NOISE_CAL with the content of (ASDM::CalDevice) noiseCal by setting numReceptor to 2 and replicating the values found in noiseCal.
Fixed a bug in the check of the size of all the elements of coupledNoiseCal. All these sizes must be >= numCalload , they were compared to numReceptor.

Re established the option --dry-run : when this option is used the filler processes all the rows of the ASDM Main table but does not write the result in the MS Main table.




git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@13901 ac9c0da5-131f-0410-9009-a9c195f183e6