New Development: Yes             JIRA Issue: No         Ready for Test: Yes      Interface Changes: Yes What Interface Changed: Some additional public methods were added to                         VisBuffer but the existing ones did not change                         so there are no required mods to existing                         code.          Test Programs: tVisibilityIteratorAsync   Put in Release Notes: No              Module(s): VisBuffer, ROVisibilityIterator, Imager, CubeSkyEquation, SkyEquation            Description:
This commital contains the asynchronous I/O functionality.  Async I/O
was introduced via the existing VisBuffer, VisibilityIterator framework.
Two new sublcasses were introduced VisBufferAsync and ROVisibilityIteratorAsync.
When the async classes are used in place of their superclasses the filling
of VisBuffers is carried out asynchronously by a new thread.  Small changes
need to be made to existing CASA code to utilize the asynchronous I/O functionality.
In addition explicit enabling entries must be made in the user's .casarc file.
Some changes were made to the VisBufferAsync and ROVisibilityIterator to make
their functionality compatible with the new subclasses.  As such it is possible
that existing code not utilizing the async i/o could still be adversely affected.

This commit also includes modifications to the Imager and CubeSkyEquation which
were slightly modified to allow them to utilize the async i/o functionality.

More information about the asychronous I/O functionality is available on 
https://safe.nrao.edu/wiki/bin/view/Software/AsyncIOon110111.
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@14083 ac9c0da5-131f-0410-9009-a9c195f183e6