New Development: Yes             JIRA Issue: No         Ready for Test: No      Interface Changes: Yes What Interface Changed: Please list interface changes      Changes in interface required to make re-samplers derived from             the base class.      A-Projection and rGridFT FTMachines are now consturcted from a             counted point of the re-sampler object.      Appropriate changes in the Imager2.cc to use the new interface of             these FTMachines.          Test Programs: List test programs   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Describe your changes here...
All the changes were to fit the various objects in the inheritance
hierarchy consistent with the design.  This required converting a few
private objects in a few classes to counted pointer, which in-turn
required changes in places where these objects were used.

The new FTMachines now use the MultiThreaded re-sampler framework.
Except rGridFT, the rest currently work only in the uni-threaded
mode. 



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active/code/synthesis@14752 ac9c0da5-131f-0410-9009-a9c195f183e6