New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes              Module(s): Module Names change impacts.            Description: Describe your changes here...
Fixed the bug that disallowed removing all baselines (cross- and
auto-correlation) involving a giving list of antennas.  Now the
negation operator works in a logically consistent manner.  To negate
(unselect), construct the equivalent expression to *select* the
baselines, and then simply apply the negation operator to it to not
have those baselines in the selected MS.  E.g., to drop all auto- and
cross-correlation baselines with, say antenna 3, the expression would
be "!(3&&*)".  The expression inside the brackets says "all auto- and
cross-correlation baselines between antenna 3 and all other
antennas".  This expression is then negated.  Note that the
parenthesis are not required, but can be used to make the expressions
more readable.
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@16162 ac9c0da5-131f-0410-9009-a9c195f183e6