New Development: No             JIRA Issue: Yes (CAS-2046)         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes              Module(s): Module Names change impacts.            Description: Describe your changes here...
When supplied an index (an integer) for field selection which is out
of range, the parser was supposed to produce an error message with a
tip "[TIP: Double-quoted strings forces name matching]".  The logic
that controlled this was erroneous and this message was never produced
(and could have helped, particularly for the case reported in
CAS-2046).

This is an intermediate but, IMO, still a fix that can be quite useful
in reducing grief for folks who, for reasons I cannot even begin to
understand, like to give names which are valid numbers (integers) as
well.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@16872 ac9c0da5-131f-0410-9009-a9c195f183e6