New Development: Yes             JIRA Issue: Yes CAS-3532.         Ready for Test: No      Interface Changes: Yes What Interface Changed: see below          Test Programs: none   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Describe your changes here...
Added new functionality to allow images to be use as Voltage pattern for ALMA antenna.
right now if nothing is done 1D rotated Airy patterns are used as before.
but if in the user's .casarc the following flag is set to true

alma.vp.image: true

Then a 2D image of the beam of the antenna is used ...right now the image is hard wired to beam responses path as defined in MeasTable + ALMA_AIRY_12M.VP  (and + ALMA_AIRY_7M.VP) for the 12m dish (and 7m dish respectively).





git-svn-id: https://svn.cv.nrao.edu/svn/casa/active/code/synthesis@17043 ac9c0da5-131f-0410-9009-a9c195f183e6