New Development: Yes             JIRA Issue: Yes         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes              Module(s): Module Names change impacts.            Description: Describe your changes here...
While ANT and STATION expressions alone may parse to a finite list of
IDs, ANT@STATION might not.  This case was not caught.  It is now.

Unfortunately, by the ANT@STATION is parsed to a single ID, ANT and
STATION expressions have already been parsed into a list of IDs.  The
error message therefore cannot produce the string tokens for ANT and
STATION IDs (at least not without caching of these tokens).  So the
error message for now containt the IDs associated with ANT and STATION
tokens.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active/casacore@17170 ac9c0da5-131f-0410-9009-a9c195f183e6