New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes              Module(s): Module Names change impacts.            Description: Describe your changes here...
By mistake, the reset() method was declared as global in .cc file and
a public method by the same name (but not with all the required
functionality) was in the .h file.  Such a code (obviously wrong -
syntactically as well as functionally), kept Jeninks and most of the
regressions happy.  Worrying.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@18463 ac9c0da5-131f-0410-9009-a9c195f183e6