New Development: No             JIRA Issue: CAS-2932       Ready to Release: Yes      Interface Changes: No What Interface Changed: None          Test Programs: importuvfits   Put in Release Notes: No              Module(s): msfits            Description: > Hi, Eric, > In the memo 117, page 10, SOURCE and FREQSEL are listed as > mandatory columns of AIPS UV table. > > In /lustre/mkrauss/helpdesk/1.48F27CB_AY0015_1986OCT06_1.uvfits, > these 2 columns are missing. > > Which one is the way to go? > > Thanks > Honglin
The memo has not differentiated a 3rd category of column - mandatory
when needed and nearly mandatory to be omitted when not needed. These
two fit into that category. If there are multiple sources, the SOURCE
column is required (and a SOURCE table must be present). If there are
not, it is a good idea to omit both. FREQSEL columns are only useful
when there are more than one FQID (more than one row in the FQ table)
and should be omitted when there are not. NOte that the FQ table is
required in all cases.

Looks like I should adjust the wording of the memo.

Thanks 



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@20380 ac9c0da5-131f-0410-9009-a9c195f183e6