New Development: No             JIRA Issue: Yes (CAS-4468)         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes              Module(s): Module Names change impacts.            Description: Describe your changes here...
The plugable error handler for antenna/baseline selection expressions
was not being reset.  As a result, a previous error was remembered
when the MSSelection object is used in a re-entrant manner.

The error hanlder object is now constructed fresh in the call to
MSSelection::toTableExprNode() and reset if it already exists.

M    MeasurementSets/MSSelectionErrorHandler.{h,cc} 
         Added the reset() method.  

M    MeasurementSets/MSSelection.cc
         Reconstruct the antenna selection error handler or reset a
         pre-existing one.


git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@20976 ac9c0da5-131f-0410-9009-a9c195f183e6