New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes              Module(s): Module Names change impacts.            Description: Describe your changes here...
PSTerm when a NOOP leaves the supplied array unmodifed.  This is
consistent with the behaviour of other CFTerms.
	
AWConvFunc.cc: PSTerm is now ON.

Imager2.cc: PSTerm is set to NoOp() for awproject.  Perhaps it will be
useful to also have a im.tool option to control this (useful for non
black-box users).



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@21633 ac9c0da5-131f-0410-9009-a9c195f183e6