New Development: Yes             JIRA Issue: Yes (CAS-4616)         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes              Module(s): Module Names change impacts.            Description: Describe your changes here...
Added the NoOpATerm class, which is what is used now if the A-term is
switched off. 	

Moved all generic functionality from specialized ATerm classes to
ATerm.cc. 

Modified createTelescopeATerm() to return NoOpATerm if required.

AWProjectWBFT.cc: Switch back on normalization by SoW.  This needs to
be revisited to work with MT-MFS.

AWProjectFT.cc: Fixed a typo in a info. log message.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@21936 ac9c0da5-131f-0410-9009-a9c195f183e6