New Development: No             JIRA Issue: Yes CAS-2822         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: all sd unit tests   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Describe your changes here...
Defined a decorator function for sd tasks that does 1) set origin 
to the logger, and 2) handle any error. All sd tasks are rewritten 
to use this decorator. 

For sdplot task, I have fixed a few bugs in worker class and updated 
unit test to fit current implementation. Note, however, that there are 
still several errors in sdplot unit test that should be fixed.
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@22051 ac9c0da5-131f-0410-9009-a9c195f183e6