New Development: Yes             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
Modified {Table,Scalar,Array}Column to allow creation of non-writable
versions of the writable objects.  This allows the objects to be
instance members of other classes (e.g., VisibilityIterator) which
might need writable or nonwritable columns.  If the xxxColumn object is
created as non-writable, it will throw a runtime exception if the write
methods are called.  

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@22129 ac9c0da5-131f-0410-9009-a9c195f183e6