New Development: Yes             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
Includes the new VI/VB framework and the start of the new TVI framework.  
Most new classes are suffixed with a "2" so they will exist in parallel
with the existing VI/VB classes.  Once phaseover to the new framework is
complete, the old VI/VB classes will disappear and the "2" suffix will
be deleted.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@22130 ac9c0da5-131f-0410-9009-a9c195f183e6