New Development: No             JIRA Issue: Yes CAS-4805         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
A while back some of the Table error messages were augmented to print out the full path of the file rather than just the filename itself.  This breaks the brittle testing method which diffs the output between the old run and the new.  I added a messsage in tTable.cc to print out that an exception occurred and then bracketed the actual error message with ">>>" and "<<<" so that the message containing the unpredictable path is ignored.  I also updated the expected output file, tTable.out, to match.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk/casacore@22552 ac9c0da5-131f-0410-9009-a9c195f183e6