New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
VisibilityIterator2 inherited the on-the-fly model data generation
features of the original.  To allow that model generation code to be
used until a new version is created to accept the current code and
adapter class, VisBuffer2Adapter was built wrap around VisBuffer2 and
allow it to perform many of the features of the original VisBuffer.
However, it turns out that at least one of the needed features was not
provided (access via the VisBuffer to the subtable columns).  This fix
adds that functionality in.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk/code/synthesis@22619 ac9c0da5-131f-0410-9009-a9c195f183e6