New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes/No              Module(s): Module Names change impacts.            Description: Describe your changes here...
Another workaround for circular dependency issue. 
I have defined an equivalent function to heuristics class that 
is used in sdtablereader module in that module. By using that 
function, circular dependency is resolved. 

The problem of this workaround is a duplication of the 
code: there are two almost same codes in hsd.heuristics and 
infrastructure.sdtablereader. But this is quite minor. 

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project/pipeline@23780 ac9c0da5-131f-0410-9009-a9c195f183e6