New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
Cleaned up the vast majority of the warning shown in a casacore build
with -Wextra and lesser warnings enabled.  Include were some cleanups
to some old Fortran code which apparently was taken from a book dated
1975; these used flow control constructs that modern day Fortran has
deprecated.  I've ran casacore's unit tests but with FFTW present, I
don't think the Fortran coded FFTs are called and are thus untested.

The few remaining warnings look to be either harmless while requiring
lots of work to change and one is a function pointer kluge that
doesn't appear to have a way to fix.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@27264 ac9c0da5-131f-0410-9009-a9c195f183e6