New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
Modified VI/VB 2 to address performance problem.  Getting the frequency involved
querying the MS about the observatory's frame of reference which turned out to 
be a very slow operation.  VI now caches the observatory's frame to reduce the
number of times it is looked up.  Sped small test code up by a factor of 10.
Other uses might not see this big an improvement.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@30599 ac9c0da5-131f-0410-9009-a9c195f183e6