New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
The VI2 was reversing the order of frequencies when the decreased across the spectral window.  There was no corresponding order change for the data which made the frequencies incorrect, and even if that had been done, it would have made it a nightmare to use.  So the frequency order reversal is removed in this commit.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@31151 ac9c0da5-131f-0410-9009-a9c195f183e6