New Development: No             JIRA Issue: cas-7176         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
Redid the tile caching logic controlled in VisibilityIteratorImpl2.cc.
This might fix the somewhat random order of magnitude slowdown
experienced by the pipeline in cas-7176 but given the random nature of
the problem, it will require a number of runs to be sure that the
problem was indeed fixed.

There is logic in the setTileCache routine that handles concatenated
files differently than normal tables.  I did not change the logic but
put in a #warning to indicate that it needs to be looked at more
closely; thus it's possible that cache-related speedups may not be
applied to concat tables by the current fix.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/release-4_3@31816 ac9c0da5-131f-0410-9009-a9c195f183e6