New Development: No             JIRA Issue: cas-7176         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
Doubled the size of the tile cache for the big data cubes.  This is to
help with problems where the autocorrelations for all DDIDs occur
first and then the corresponding cross correlations; this can cause a
single VisBuffer fill to require data from two sets of tiles which
ends up causing reads (or system disk cache accesses) at a high rate.
The proper solution is to modify the filler so that auto correlations
and cross correlations are located consecutively.  Another workaround
is to exclude, by selection, the autocorrelations if they aren't
needed for the task at hand.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/release-4_3@31831 ac9c0da5-131f-0410-9009-a9c195f183e6