New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s): Build            Description:
The previous commit to this file only fixed one case where the
Qt4Macros.cmake mods affected CASA's CASA.cmake file.  Added
cmake-version triggered logic that will use the appropriate
qt4_extract_option signature.  If the old signature is used the
symptom will be an undefined reference to qInitResources_XXX. If the
new signature is used with an older version of qt4_extract_options it
might be harmless(?).

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@32106 ac9c0da5-131f-0410-9009-a9c195f183e6