New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
Modified the "reform" method of array so that you can specify that the
reform is acceptable providing that the new shape stays within the
size of the underlying storage.  Before this mod, the reshaping had to
use exactly the same number of elements.  This mod allows an array to
change between different shapes without requring further reallocation.
Probably a bit of a kluge, but helps in cases where a large array is
changing shapes frequently.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk/casacore@32938 ac9c0da5-131f-0410-9009-a9c195f183e6