only parallelize quicksort when there is significant work to do
gnu openmp will always start a full group of threads regardless of loop
iteration count. So e.g. on a 16 core machine two threads are working
and the rest will wait in the barrier which is a spinlock with approx
3ms spin time.
When the machine is oversubscribed this causes massive slowdowns as the
threads holding the logs need to be scheduled before the parallel block
can end.

To avoid this issue only attempt parallization when there is enough work
to do for the two threads. Alternatively one could also limit the number
of threads to min(2, OMP_NUM_THREADS) or implement more efficient omp task
based parallization which has much smaller spincounts in the
barriers.

cherrypick of gh b940683368ff776eb8001af63f7f85502b75220f

git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@33077 ac9c0da5-131f-0410-9009-a9c195f183e6