New Development: No             JIRA Issue: cas-7491         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
Modified VI2 framework so that when averaging the requested
frequencies are reported correctly.  When averaging the VI2 is often
looking at a segment of the MS that is ahead of the VisBuffer2 just
produced but the frequency access logic assumes that the VI2 and the
VB2 are in sync.  Changed the logic so that when the VB2 requests the
frequency from the VI2 it provides a spectral window ID and an MS ID
in addition to the time.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/release-4_4@33201 ac9c0da5-131f-0410-9009-a9c195f183e6