New Development: No             JIRA Issue: cas-7601         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
Added two new methods to VisBuffer2 which will carry along the
correlation information as a component of the VisBuffer2:
getCorrelationTypesDefined and getCorrelationTypesSelected.  Both
return Vector<Stokes::StokesTypes>.  The "defined" method returns the
set of correlation types contained in the data while the "selected"
method returns only those correlation types that were actually
selected (and in the same order).

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/release-4_4@33497 ac9c0da5-131f-0410-9009-a9c195f183e6