New Development: No             JIRA Issue: cas-7705         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
Modified VI2 to allow the user to select, via casarc, whether to use
the existing tile cache sizing logic or the built-in casacore logic.  The
setting is:

VisibilityIterator2.AutoTileCacheSizing: False/True

Setting it to false (the current default) leaves things as they currently are
(i.e., the VI2 sets the tile cache size); if True then the casacore logic
is allowed to automatically size the tile cache.  



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@34322 ac9c0da5-131f-0410-9009-a9c195f183e6