New Development: Yes             JIRA Issue: Yes CAS-7751, CAS-7767         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes              Module(s): Module Names change impacts.            Description: Describe your changes here...
Handling of errors in SPW parser is via an error handler now. This
brings uniform error reporting and allows changing the error reporting
behaviour programmatically by the clients.

Instead of throwing an exception for no-match for some components of
the expression, the error is now reported as a warning.  If no
components were matched, an exception is thrown.

Added new class MSSSpwErrorHandler.
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/release-4_5@34481 ac9c0da5-131f-0410-9009-a9c195f183e6