New Development: Yes             JIRA Issue: No         Ready for Test:      Interface Changes: What Interface Changed:          Test Programs:   Put in Release Notes:              Module(s):            Description:
Reworked VI2 framework to decouple VB2s, except for the last one in a
transformation chain, from the VI2.  Instead, they'll reference the
ViImplmentation2 class which will provide any needed functionality.
The last VB2 will contain a link to the VI2 but only as a convenience
to VI2 clients (mostly to get at meta data).

Stripped most of the protected data access methods out of the VI2
since the VB2s won't be using them.  This greatly simplifies the
class since it was simply passing things through to the underlying
VII2.
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@35556 ac9c0da5-131f-0410-9009-a9c195f183e6