New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
Small fix to address a problem encountered by Sanjay.  When setting an
array-valued VisBuffer2 component to a scalar, an exception is thrown
if the component hasn't been read (it has no shape).  The fix will
detect this situation and provide an uninitialized array with the
appropriate shape before the setting is allowed to proceed (and
succeed).
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@36627 ac9c0da5-131f-0410-9009-a9c195f183e6