New Development: No             JIRA Issue: cas-6119         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes:              Module(s):            Description:
Modified logic used to generate SWIG tool bindings to remove leak.
The bindings tended to leak on most of the calls to CASA tools.  The
major change was to modify the "code generation" logic in casa_typemaps.i
so that when objects were allocated with "new" a function-scope unique_ptr
was added to delete the object when the function exited.  A small mod
was made casa2c++h.xsl which called the initialize_variant method which
leaked.  A copy constructor and assignment operator were added for
the variant datatype to allow leak-free copying.
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@37068 ac9c0da5-131f-0410-9009-a9c195f183e6