New Development: Yes             JIRA Issue: cas-8438         Ready for Test: Yes      Interface Changes: What Interface Changed:          Test Programs:   Put in Release Notes:              Module(s):            Description:
Flushed in change that uses new casacore functionality to add model and/or
corrected scratch columns that conform to the tiling shapes of the underlying
data column.  Thus if the data column is tile "optimally" then the scratch
columns will be as well.  This fixes a problem reported by Kana where
adding the scratch columns exploed a 18 MB file to an 8 GB file.  

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@37212 ac9c0da5-131f-0410-9009-a9c195f183e6